move actionTypes to types, refactor types and actionTypes
