+++ store/reducer.tsx
@@ -1,10 +1,8 @@
-import * as actionTypes from './actionTypes';
-
 import {
   HIGHLIGHTED_RESTAURANTS,
   DEFAULT_BREADCRUMB,
 } from '../helpers/staticData';
-import { Restaurant, BreadcrumbsData } from '../helpers/types';
+import { Restaurant, BreadcrumbsData, Actions } from '../helpers/types';
 
 export type InitialState = {
   favorites: Restaurant[];
@@ -18,29 +16,29 @@ const initialState: InitialState = {
   breadcrumbs: [DEFAULT_BREADCRUMB],
 };
 
-const reducer = (state = initialState, action: actionTypes.Actions) => {
+const reducer = (state = initialState, action: Actions) => {
   switch (action.type) {
-    case actionTypes.SET_FAVORITE:
+    case 'SET_FAVORITE':
       return {
         ...state,
         favorites: [...state.favorites, action.favorite],
       };
-    case actionTypes.SET_RELATED_RESTAURANTS:
+    case 'SET_RELATED_RESTAURANTS':
       return {
         ...state,
         relatedRestaurants: action.relatedRestaurants,
       };
-    case actionTypes.CLEAR_RELATED_RESTAURANTS:
+    case 'CLEAR_RELATED_RESTAURANTS':
       return {
         ...state,
         relatedRestaurants: initialState.relatedRestaurants,
       };
-    case actionTypes.ADD_BREADCRUMBS:
+    case 'ADD_BREADCRUMBS':
       return {
         ...state,
         breadcrumbs: [...state.breadcrumbs, action.breadcrumbs],
       };
-    case actionTypes.REPLACE_BREADCRUMBS:
+    case 'REPLACE_BREADCRUMBS':
       return {
         ...state,
         breadcrumbs: [
